input_driver.state 返値を32bit化する危険な試み

+++ input/drivers/android_input.c
@@ -1360,7 +1360,7 @@ bool android_run_events(void *data)
    return true;
 }
 
-static int16_t android_input_state(
+static int32_t android_input_state(
       void *data,
       const input_device_driver_t *joypad,
       const input_device_driver_t *sec_joypad,
@@ -1380,7 +1380,7 @@ static int16_t android_input_state(
          if (id == RETRO_DEVICE_ID_JOYPAD_MASK)
          {
             unsigned i;
-            int16_t ret = 0;
+            int32_t ret = 0;
             for (i = 0; i < RARCH_FIRST_CUSTOM_BIND; i++)
             {
                if (binds[port][i].valid)
